Further Details on OTRLSchema.pdf
Explicitly stating what is a primary key and what is a foreign key.  Thought process is an Employer Refers an Employee to the referral form. An Employee sends the link to the form and the Employer fills out the form and submits it. This is why the Referral form has a submission time and the score  of the form in the table.  While designing we thought about the queries that would be made, and decided to go with this concept since the queries would perform better and be less impactful to DynamoDB.