Specifications on how lambda will be integrated
Algorithms are included in the document with how the reputation score of the survey will be calculated and the reputation score of the employee will be calculated. This takes into consideration that an employee might have multiple people submitting the a referral form for them. The weights of each question will be stored in Lambda and be utilized when needed.